requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dj-database-url==1.0.0
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.5
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.5
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.5
urllib3==1.26.12
